Move appropriate parts of At, Cons, Each and Empty to optics-core
