+++ optics-th/src/Optics/TH/Internal/Product.hs
@@ -30,7 +30,7 @@ import Data.Either.Optics
 import Data.Tuple.Optics
 import Data.Set.Optics
 import Language.Haskell.TH.Optics
-import Optics.Core
+import Optics.Core hiding (cons)
 import Optics.TH.Internal.Utils
 
 ------------------------------------------------------------------------
